requirements.txt
earthengine-api>=0.1.380
geemap>=0.30.0
pandas>=2.0.0
geopandas>=0.14.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
rasterio>=1.3.0
folium>=0.14.0
python-dotenv>=1.0.0
plotly>=5.17.0
shapely>=2.0.0
pyproj>=3.6.0
